@@ -26,25 +26,27 @@
     "source": "https://github.com/BugBuster1701/contao-dlstats-statistic-export-bundle"
   },
   "config": {
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "contao-components/installer": true,
+      "contao/manager-plugin": true,
+      "php-http/discovery": true
+    }
   },
   "require": {
-    "php": "^7.4 || ^8.0",
+    "php": "^8.1",
     "ext-pdo": "*",
-    "bugbuster/contao-dlstats-bundle": "^1.1",
-    "contao/core-bundle": "^4.9",
-    "composer/semver": "^3.0",
-    "doctrine/dbal": "^2.13 || ^3.3",
-    "jean85/pretty-package-versions": "^1.5",
+    "bugbuster/contao-dlstats-bundle": "^1.4",
+    "contao/core-bundle": "^5.2",
+    "doctrine/dbal": "^3.6",
     "phpoffice/phpspreadsheet": "^1.16",
-    "symfony/dependency-injection": "^4.3 || ^5.0",
-    "symfony/form": "^4.3 || ^5.0",
-    "symfony/translation": "^4.3 || ^5.0 ",
-    "symfony/validator": "^4.3 || ^5.0"
+    "symfony/dependency-injection": "^6.4",
+    "symfony/form": "^6.4",
+    "symfony/translation": "^6.4 ",
+    "symfony/validator": "^6.4"
   },
   "require-dev": {
-    "contao/manager-plugin": "^2.1",
-    "friendsofphp/php-cs-fixer": "^2.12"
+    "contao/manager-plugin": "^2.1"
   },
   "autoload": {
     "psr-4": {
@@ -53,5 +55,17 @@
   },
   "extra": {
     "contao-manager-plugin": "BugBuster\\DlstatsExportBundle\\ContaoManager\\Plugin"
+  },
+  "scripts": {
+    "all": [
+        "@cs-fixer"
+    ],
+    "cs-fixer": [
+        "@ecs",
+        "@ecs-legacy"
+    ],
+    "ecs": "@php tools/ecs/vendor/bin/ecs check src/ --config tools/ecs/config/default.php --fix --ansi",
+    "ecs-legacy": "@php tools/ecs/vendor/bin/ecs check src/Resources/contao/ --config tools/ecs/config/legacy.php --fix --ansi",
+    "tx-pull": "tx pull -a --minimum-perc 10"
   }
 }
